+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/audio/wav/RiffChunk.kt
@@ -1,4 +1,4 @@
-package org.wycliffeassociates.otter.common.wav
+package org.wycliffeassociates.otter.common.audio.wav
 
 import java.nio.ByteBuffer
