+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/audio/wav/WavMetadata.kt
@@ -1,4 +1,4 @@
-package org.wycliffeassociates.otter.common.wav
+package org.wycliffeassociates.otter.common.audio.wav
 
 import java.io.OutputStream
 import java.nio.ByteBuffer
