+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/recorder/ActiveRecordingRenderer.kt
@@ -3,8 +3,7 @@ package org.wycliffeassociates.otter.common.recorder
 import io.reactivex.Observable
 import io.reactivex.schedulers.Schedulers
 import org.wycliffeassociates.otter.common.collections.FloatRingBuffer
-import org.wycliffeassociates.otter.common.wav.DEFAULT_SAMPLE_RATE
-import org.wycliffeassociates.otter.common.wav.WavFile
+import org.wycliffeassociates.otter.common.audio.wav.DEFAULT_SAMPLE_RATE
 import java.nio.ByteBuffer
 import java.nio.ByteOrder
